created new menuGroups component
